+++ realm/realm-jni/src/io_realm_internal_SharedGroup.cpp
@@ -324,3 +324,22 @@ JNIEXPORT jlongArray JNICALL Java_io_realm_internal_SharedGroup_nativeGetVersion
 
     return version_data;
 }
+
+JNIEXPORT jboolean JNICALL Java_io_realm_internal_SharedGroup_nativeWaitForChange
+  (JNIEnv *env, jobject, jlong native_ptr)
+{
+    TR_ENTER_PTR(native_ptr)
+    try {
+        return static_cast<jboolean>(SG(native_ptr)->wait_for_change());
+    } CATCH_STD()
+    return false;
+}
+
+JNIEXPORT void JNICALL Java_io_realm_internal_SharedGroup_nativeStopWaitForChange
+  (JNIEnv *env, jobject, jlong native_ptr)
+{
+    TR_ENTER_PTR(native_ptr)
+    try {
+        SG(native_ptr)->wait_for_change_release();
+    } CATCH_STD()
+}